Implement full ShareKey authorization for Blob Service
The particular problem we had to solve here is to enable the SharedKey
filter to have access to the "Content-Length" of the request, since
that header is needed for authorizing the request.

Since the "Content-Length" of the request is known only very late in the
pipeline, so we had to introduce a new listener interface
(EntityStreamingListener) that filters can register into.

Filters then get called back later in the pipeline, just before starting
sending bytes on the connection. This allows them to access all the
headers (and also gives them a last chance to update them).