Utilize factory method pattern instead of class. fixes #99
class XxxService should have two methods:

static XxxContract create();
static XxxContract create(Configuration config);

As a class it will not implement the XxxContract interface.

As a class it will not have public constructors.

To avoid confusion it will have a private default
constructor (this is the Java way of doing "static class")