[BUGFIX] - Fixing up and commenting out tests
* For JobOperationsTests, I found a server side bug while testing this
call in the .NET SDK, so I was not surprised to see that it fails here
as well.

* For CatalogOperationsTests, I commented out "BUCKETS" from the script
because apparently that keyword is deprecated.  Furthermore, I commented
out the "canCreateUpdateDeleteSecretsAndCredentials" test because the
scripts that are being used are deprecated.  According to the error
message, I must use Azure Powershell cmdlets for such opeartions.  I am
not surprised at this because, to my knowledge, it's been a while since
this test case was run.  I will look into how I can test this in the
near future.

* Other than that, I made sure that JobOperationsTests as well as
AccountOperationsTests (for both ADLA and ADLS) passed.