Add test for XML serialization in RestProxy
Note that this highlights a significant generator issue-- it's easy to consume the XML received from a service, but hard to produce it in a form we know will be acceptable to the server. The Java object we created to represent the argument to this method has <SingletonList> as its root element.