Appveyor detection fix (#358)
EPH test cases which need actual Azure entities look in environment variables for the connection strings and skip if not found. Those environment variables shouldn't be set in Appveyor but the test cases were trying to run anyway, and failing. Syntax for appveyor.yml allows for setting environment variables but it isn't obvious how to delete them, so change code to also check APPVEYOR environment variable.