Introduce two build profiles into the client pom.xml (#2939)
Introduce two build profiles into the client pom.xml, one for building with a language target of JDK 8 (which is the default), and one to build targeted at the LTS version of Java - currently configured to be Java 11. Activate the LTS target with `-Djava-lts`.

Also fix URL issue in JavaDoc output, exclude module directories being part of the link structure in JDK 9+ builds of JavaDoc, and remove shade plugin as it is unused at present.