Vijay long poll link abort fix (#7995)
When a long poll receive request (something that waits for 60 minutes) is pending and the link is aborted, then that pending request will remain pending for the whole timeout period even though messages are available in the entity. Fixed it. Added a test. Also updating the version of the SDK.

Also, Changing two acceptAsync methods to use internal thread pool.
And depcreating and renaming two methods as their spellings were wrong.