Issue 8079: Non thread safe call to write blob operation.
Enabling api rest calls using the TracerProxy to run in multi thread a environment. This doesn't consider changes in classloader (like adding a new jar containing a new implementation of Tracer), but at least works for threads running in parallel.

Needed to converted to a singleton because lock mechanisms in lazy load for static fields is not the best practise. The lazy load could be removed, but depending on the order/priority for loading classes, we could introduce another bug.

To avoid any problem, a singleton with lazy load on the list of Tracer providers was introduced.