Event Hubs track 1 receive watchdog (#10197)
Adds a watchdog to the receive path. When enabled, if all receivers created from a given EventHubClient go silent (zero traffic) for a user-specified period, the watchdog assumes that there is a problem with the underlying connection and recreates it. This change is directed at specific customers who have very high traffic and never expect to have, for example, even one minute with legitimately no traffic. This is a setting that needs to be public but which most users will never need and should never touch.