Add Model API tests to ADT
Fixed a few test assets as well. While they were accepted by the service when the context field was just one value, the service always returns a model such that the context is an array with one value instead. This makes it easier to test that models retrieved from the service match our expectations