Adding a comment to reference Hooks#KEY_ON_ERROR_DROPPED (#15365)
* Stop onErrorDropped events from leaking into default Reactor Hook

There is a race condition between cancellation (via Reactor timeout) and completing the request in the Rntbd stack (via error/request timeout or successfully) which right now results in onErrorDropped events being raised to the default Reactor Hook. This results in ERROR level logging - functionally it is harmless (we do proper clean-up and functionally it is expected that the cancellation takes precedence and we ignore any subsequent success/transport exception). To avoid confusing customers I have injected a local hook now that will just log these errors with a DEBUG level - hopefully that will reduce the noise level for customers and stop causing confusion.

* Reacting to code review feedback

* Avoiding that we expire 1 ns too late :-)

* Making reactorContext final

* Adding a comment to reference Hooks#KEY_ON_ERROR_DROPPED