Allow for custom serializer in ADT, add mappers for event route and incoming relationship types (#15431)
We shouldn't expose a generated type like EventRoute, EventRouteListOptions, and IncomingRelationship if those types come with a validate() function. Instead, I have created wrapper classes for each so that we don't expose the validate() function while still getting the client side validation that we want.

For the json serializer, the changes are modeled after the SearchClient for Azure Search. Like in that client, the custom serializer does not get passed down into the protocol layer, but is used anywhere there is json serialization/deserialization within the convenience layer