Use the builder to create table clients from service clients (#16320)
* Always use the builder to make clients

Just re-using the same implementation won't work because the table
client needs to use a different serializerAdapter. This still re-uses
the same pipeline (incl. HTTP client) so it shouldn't cause any perf
regression.

* Correctly handle TokenCredentials in builders

Previously, TablesSharedKeyCredential was a TokenCredential, and the
builders always assumed that any TokenCredential the customer passed was
a shared key credential. This behavior was broken and is now corrected.

* Add test to cover bug #16292