Fixing a bug in management client that is sending atom xml elements in enity description (#17046)
out of order in an update call. Because of this bug, deserializer on the service ignores
these out of order elements, and persists elements with default values in the righr order
resulting in these out of order elements getting duplicated in the resource xml.
This fix sends xml elements in the correct order in update call.