RemoteRendering: Handle nulls in session properties (#19582)
The service changed recently and is no longer returning properties with null values. This exposed an assumption in the client which had to be fixed. I've modified the session records so this fact is tested in playback.

The change in the service and the fact that the environment variables were wrongly named, suggest that the test pipeline has not been running tests live. Otherwise I cannot explain how the test pipeline has been green recently.