move to spark 3.1.1 fixed schema inference issue on write (#19651)
move to spark 3.1.1 
Due to a bug in spark 3.0.x schema-inference used to be invoked even for write operation. after moving to spark 3.1.1 this is resolved (I verified this).

I still kept `ACCEPT_ANY_SCHEMA` as a capability @ealsur  do you see any reason `ACCEPT_ANY_SCHEMA` need to be removed?

TODO: we should change the folder structure to reflect the artifactId change `sdk/cosmos/azure-cosmos-spark_3-0_2-12` -> `sdk/cosmos/azure-cosmos-spark_3-1_2-12`. I am not going to do that for now because it will create unnecessary git merge issues for others working on the project (we will do later).