Synchronizing populatePropertyBag as additional fallback to avoid transient invalid json in Changefeed state (#21020)
It is not fully clear yet how we occasionally end-up with invalid json for the ChangeFeedStartFrom settings (json would have "Type": "Now", "Type": "Now") - the populate property bag method uses Map<>.put (LinkedHashMap) to add the property - which should never result in duplicates. Synchronizing this part is just an attempt to reduce risk - not because there is a known multi-threaded usage that could result in the duplicates.