CosmosDB: ChangeFeedProcessor fix for ownership race condition (#25376)
* CosmosDB: ChangeFeedProcessor fix for ownership race condition
The issue occurs when the current worker is too slow at processing or renewing the current lease, and gives an opportunity to another CFP instance to take over. If that second worker encounter any errors while processing the feeds, the ownership will be set to 'null'. When the first worker finally resumes, it will continue to checkpoint and renew the lease because the current check which establishes ownership missing the 'null' owner case.