Session token optimization and ChangeFeedProcessor bug fixes
By default we don't scope the global session token in Gateway when partition key ID is part of the request. For containers with a very large number of partitions and multi-region account setting, this can result in "Request headers too long" errors.
The change is to look first at the current request's headers and if the request is scoped to one specific partition, we should also scope the session token to that specific partition.

ChangeFeedProcessor related changes: when the renewer task is cancelled, we also need to close the processor task even if we are about to check-point the current state. The cange also ensures that any exceptions are prioritized accordinally.

Added more logs to CFP to allow for better monitoring of the current processing.

Added test case for the scoped session token and simplified the existing CFP test.