CosmosDB: Add implementation for CosmosAuthorizationTokenResolver (#26118)
* Add implementation for CosmosAuthorizationTokenResolver

This interface is for client side implementation, which can be used for initializing CosmosAsyncClient without passing master key, resource token and permission feed.
Each time the SDK create request for CosmosDB, authorization token is generated based on that request at client side which enables creation of one CosmosAsyncClient per application shared across various users with different resource permissions.
The new interface provides a mechanism for the client application to pass in the value of the authorization header to be used by the service side when authenticating and authorizing each request. This value can be a resource token, an AAD token, any value that the client application logic would construct based on the type of request to be sent out.