CosmosDB: ChangeFeedProcessor fix for check-pointing race condition (#26143)
* ChangeFeedProcessor related changes
When the renewer task is cancelled, we also need to close the processor task even if we are about to check-point the current state. The change also ensures that any exceptions are prioritized accordingly.
Added more logs to CFP to allow for better monitoring of the current processing. Improve logging when unused or expired leases are detected.
Update CFP test in order to simplify the logic.