AzureSeekableByteChannel - permit no-op seeks (to current position) in write mode
This change does not enable random-access nor forward-only seeking on writable
channels, but it makes the channel more permissive and Java API compatible when
used by (Azure-unaware) code that may call `location(long)` on such a channel.
As long as the position is not actually changing, the call does no harm.

As a result of this change, calling `position(n)` where `n < 0` on an open
channel now throws IllegalArgumentException per the Java API specification,
regardless of whether the channel is readable or writable.  Previously, a
writeable channel would throw an exception due to its non-readable status.