CosmosDB: ChangeFeedProcessor custom scheduler and other fixes (#26750)
* Introduce beta API to set custom Reactor scheduler to be used by the ChangeFeedProcessor implementation.
Fix an issue related to leases that were found expired. We should ensure that current CFP instance will pick at least one lease in such cases (exception is if maximum scale count was reached), even when it might break the equal partitioning.

* Update logs (and add more) in order to get a better understanding of when things don't go as expected.
Add a check in the acquiring of a lease to ensure that ownership did not change in between the updates.
Remove "lock" that was not needed anymore; we rate limit the upstream instead.