Require test agent phone number in livetests
Throw an exception when there is no phone number assigned to the current test agent. This only applies when not skipping update capabilities live tests.

The main purpose of this is for the test run to fail when it isn't properly configured, instead of silently falling back on the default method for setting the test phone number.