[Communication] - phone-numbers - Implement a phone number pool for live tests (#27072)
* Implement a phone number pool for live tests

* Fix pipeline script

* Add MatrixConfigs param to communication template

This allows for individual packages to customize their job matrix if needed.

* Require test agent phone number in livetests

Throw an exception when there is no phone number assigned to the current test agent. This only applies when not skipping update capabilities live tests.

The main purpose of this is for the test run to fail when it isn't properly configured, instead of silently falling back on the default method for setting the test phone number.