Update_version.py script could throw an exception if a dependency is not present in external_dependencies file.
In the first attempt I tried to do a graceful check of the dictionary. However, that was not sufficient since that led to the correct version not being found and hence the update script simply ommited the version.
In the second attempt I updated to not throw however the root of the problem still remained.
In the current attempt I have updated the external dependencies file with the new dependencies that we have added in the system.
I have also updated the script to redeclare the newline since we could get into a state where newline was not declared and so depending on whether the newline string was found or not get incorrect updates to the file.