Enable requestBody to read buffer as duplicate. Also update OkHttpClient to not follow redirects by default. (#27960)
OkHttpClient by default has redirectfollow set to true, which makes it difficult for us to invoke our own redirect policies. Updated that. I expect this not to be a breaking change since any SDK would have had to add redirect policy for Netty client which would mean OkHttp would continue to work for them. There is a small chance that an SDK may only support OkHttpClient and so may get affected however the right fix would be to update the redirect policy there. Let me know if anyone disagrees.