Offer option to perform operations with timeout
See https://github.com/Azure/azure-sdk-for-java/issues/26465#issuecomment-1413722753 for the details of the issue, in short we have been observing regularly that some threads in a concurrentConsumer setup would stop to process further messages. A threadump shows them hanging in the complete section. After forking the repository and doing a quick test this issue can indeed be resolved by performing the .complete() action with a timeout after which the thread would recover. This change is created in a non invasive manor that offers clients the option to call the method with a given timeout while not changing any current default behavior