Re-enable test proxy unit tests
The failure was caused by the fact that these tests effectively override the `AZURE_TEST_MODE` variable so they can test record and playback in the same session. The mechanism for doing so is later than when the test proxy started, so the sequence of events is:

```
AZURE_TEST_MODE=LIVE
skip starting test proxy
AZURE_TEST_MODE=PLAYBACK (within a test)
try to talk to test proxy.
```

This isn't a concern for real tests, as they will have a consistent mode throughout.