Change expected exception for additional tenant tests (#35498)
The current exception only works when the user isn't logged into `az`. Change it so it works in both cases. (In either case we don't care about what az does, we want to test the flow building the commandline to the test.) This is worth redoing at some point with good mocks around `az`/`azd`, but for now this will quiet a noisy test.