Redact error output
A bug was reported to me in which I observed that the exception message
for the CLI process timing out can include a warning and then the access
token. Azure CLI defaults to printing warnings. The fact that the output
is not pure JSON causes us to fail parsing it, and the exception
includes the token.