[OpenAI] `prompt_annotations` and `prompt_filter_results` rename (#36908)
On the service side, there was a rename from prompt_annotations -> prompt_filter_results. This breaking change was meant for models Completions and ChatCompletions. In order to move forward with the release we commented out the impacted tests in this PR.

The changes introduced in this PR address the rename and provide a path forward mitigating as much as possible the impact of the rename.

Added backing field for prompt_annotations to be maintained during the transition period of the renaming.
prompt_annotations is only used for deserialization and should not be visible to the user
This field is to be removed once the rename change is propagated and committed to both the Completions and ChatCompletions models. (Currently, the rename is only effective in ChatCompletions)