Always use playback poll interval during playback (#41313)
Updates tests in Translation to set the poll interval using the TestBase helper method. During playback tests this will shrink the poll interval considerably as we aren't making calls to the service and no tests should be depending on how fast polling happens (as that would likely result in many race conditions).