Removing VM size check from Azure Remote Rendering sdk tests
The check is removed as we currently load-balance the session requests for the testing account specifically to avoid customer impact with regards to regular session availability. So in essence, the sessions the tests will get are not guaranteed to be 'standard' sized. Consequently, there is no point in testing this particular session property.