Internal exceptions cause havoc for App Insights - Correct whitespace
------------------------------------------------

The problem is with the Azure BlobServiceClient.createBlobContainerIfNotExists(String containerName) library method.
•	This method returns an access object for the named container.  If the container does not exist then it will be created.
o	This is implemented using an internal call to the BlobContainerClient.createBlobContainer(String containerName) method.
o	This call then uses a lower-level web service call that attempts to create the container.  This call is implemented to accept only status code 201 (container created) as a valid response.  Any other status code causes an exception to be thrown.
o	If the exception was thrown because of receiving code 409 (container exists) then the exception is squashed and the createBlobContainerIfNotExists() method returns normally.
o	Internally the exception is reported to App Insights as a hard error.  This causes havoc for anyone using App Insights.

My solution is to modify the existing Azure BlobServiceClient.createBlobContainerIfNotExists(String containerName) library method.
•	This method returns an access object for the named container.  If the container does not exist then it will be created.
o	This is implemented using a new lower-level web service call that attempts to create the container.  This call is implemented to accept status codes 201 and 409 as valid responses.  Any other status code causes an exception to be thrown.
o	If either 201 or 409 was received then the createBlobContainerIfNotExists() method returns normally.

A similar problem exists for the BlobServiceClient.deleteBlobContainerIfExists (String containerName) library method.
A similar solution is provided.