Update OpenAIClientBuilder.java to be able to handle all non-azure endpoints
Currently the only way to acquire a NonAzureOpenAIClient, which can interact with any endpoint designed to take OpenAI-api-like HTTP requests, is to pass no endpoint at all or to pass an endpoint starting with "https://api.openai.com/". An option to specify whether you want a NonAzure client or not would be appreciated.